player triggers tile event
player checks each tile that is being stepped on.
when tile is stepped on, that tile informs its parent tilemap
the tilemap informs all other tiles in the same tilemap as well as its parent area
the area invokes an event that lets the CreatureManager know that this area is ready for an creature to spawn.
The creature manager gets a tile from the area that triggered the spawn event which is to be used to get a position to spawn a creature.